requirements.txt
flask==2.3.3
python-docx==0.8.11
pandas==1.5.3
numpy==1.24.3
PyPDF2==3.0.1
openpyxl==3.1.2
markupsafe==2.1.3
pillow==10.0.0
gunicorn==21.2.0
python-dotenv==1.0.0 